% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_stats.R
\name{ave_control_centrality}
\alias{ave_control_centrality}
\title{Average Control Centrality}
\usage{
ave_control_centrality(A)
}
\arguments{
\item{A}{An n by n matrix.}
}
\value{
A length n vector of average control centrality measures \insertCite{pasqualettiControllabilityMetricsLimitations2014a}{netcontrol}, representing the overall average control of each node in the system.
}
\description{
Calculates the average control centrality of a system defined by \eqn{x_(t+1) = Ax_(t) + Bu_(t)}.
}
\examples{
A = matrix(c(0,-3,-2,2,-2,1,-1,2,-1), 3,3)

ave_control_centrality(A)
}
\references{
\insertRef{pasqualettiControllabilityMetricsLimitations2014a}{netcontrol}
}
