% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_traj.R
\name{control_step}
\alias{control_step}
\title{Compute a single timestep of a linear time invariant system with white noise disturbances and measurement error.}
\usage{
control_step(x, A, B, u, theta = NA, gamma = NA)
}
\arguments{
\item{x}{x_t, a n length vector of observations}

\item{A}{An n by n matrix of system coefficients}

\item{B}{An n by matrix of control-to-node mapping coefficients}

\item{u}{An m length vector of control inputs}

\item{theta}{A positive definite covariance matrix for white noise disturbances}

\item{gamma}{A positive definite covariance matrix for white noise measurement error}
}
\value{
A list with x: a vector of the states at t+1, and y: a vector of the observations at t+1
}
\description{
Compute a single timestep of a linear time invariant system with white noise disturbances and measurement error.
}
