% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_traj.R
\name{control_traj}
\alias{control_traj}
\title{Simulate the trajectory of a linear time invariant system with white noise disturbances and measurement error, subject to control inputs.}
\usage{
control_traj(
  t_max,
  x_0,
  A,
  B,
  theta,
  gamma,
  G_seq,
  J_func,
  S,
  Q_seq,
  R_seq,
  delta = NA,
  d_nosign = F,
  d_toggle = F
)
}
\arguments{
\item{t_max}{The number of time points to simulate.}

\item{x_0}{The starting values for the state vector.}

\item{B}{An n by matrix of control-to-node mapping coefficients}

\item{theta}{A positive definite covariance matrix for white noise disturbances}

\item{gamma}{A positive definite covariance matrix for white noise measurement error}

\item{G_seq}{A single Kalman gain matrix, or a sequence of Kalman gain matrices.}

\item{J_func}{A quadratic regulator cost function.}

\item{S}{The S matrix.}

\item{Q_seq}{A single Q matrix or a sequence of Q matrices}

\item{R_seq}{A single R matrix of a sequence of R matrices}

\item{delta}{For saturated inputs, a length 2 vector of the threshold at which the input is saturated, and the value of a saturated input. NA for no saturated inputs}

\item{d_nosign}{Logical, if inputs are saturated, should the saturation only occur for positive inputs.}

\item{d_toggle}{Logical, if inputs are saturated, should the input be a step function. (0 unless it exceeds the saturation point.)}

\item{u}{An m length vector of control inputs}
}
\value{
A list with state values, observation values, control input values and cost function values for 1:t_max timepoints.
}
\description{
Simulate the trajectory of a linear time invariant system with white noise disturbances and measurement error, subject to control inputs.
}
