% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlyap.R
\name{.dlyap_native}
\alias{.dlyap_native}
\title{Discrete Lyapunov Equation Solver, base R implementation}
\usage{
.dlyap_native(A, W)
}
\arguments{
\item{A}{}

\item{W}{}
}
\description{
Discrete Lyapunov Equation Solver, base R implementation
}
