% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_stats.R
\name{inv_average_control}
\alias{inv_average_control}
\title{Trace of the Inverse Gramian}
\usage{
inv_average_control(A, B)
}
\arguments{
\item{A}{A \eqn{n x n} matrix.}

\item{B}{A \eqn{n x m} matrix.}
}
\value{
Trace of the inverse infinite time Gramian.
}
\description{
A commonly used measure of the overall controllability of a system defined by \eqn{x_(t+1) = Ax_(t) + Bu_(t)}.
}
\examples{
A = matrix(c(0,-3,-2,2,-2,1,-1,2,-1), 3,3)
B = diag(3)

inv_average_control(A, B)
}
