% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_stats.R
\name{modal_control}
\alias{modal_control}
\title{Modal Control}
\usage{
modal_control(A, B)
}
\arguments{
\item{A}{An n by n matrix.}

\item{B}{An n by m matrix.}
}
\value{
An m by n matrix representing the control of the nth mode by the mth control input.
}
\description{
Calculates the modal control of a system defined by \eqn{x_(t+1) = Ax_(t) + Bu_(t)}.
}
\examples{
A = matrix(c(0,-3,-2,2,-2,1,-1,2,-1), 3,3)
B = diag(3)

modal_control(A, B)
}
