% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_stats.R
\name{modal_control_centrality}
\alias{modal_control_centrality}
\title{Modal Control Centrality}
\usage{
modal_control_centrality(A)
}
\arguments{
\item{A}{An n by n matrix.}
}
\value{
A length n vector of modal control centrality measures, representing the overall modal control of each node in the system.
}
\description{
Calculates the modal control centrality of a system defined by \eqn{x_(t+1) = Ax_(t)}.
}
\examples{
A = matrix(c(0,-3,-2,2,-2,1,-1,2,-1), 3,3)

modal_control_centrality(A, B)
}
