% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_traj.R
\name{S_seq_calc}
\alias{S_seq_calc}
\title{Recursively solve the Riccati equations}
\usage{
S_seq_calc(A, S, B, Q_seq, R_seq, tmax)
}
\arguments{
\item{A}{}

\item{S}{}

\item{B}{}

\item{Q_seq}{}

\item{R_seq}{}

\item{tmax}{}
}
\value{

}
\description{
Recursively solve the Riccati equations
}
